Don't proceed if watched dir isn't created
If the directory is created *after* we add an inotify watch then we
don't receive any events. We instruct add_watch method of class
WatchManager to raise an exception in this case. This allows us to
catch it and retry indefinitely to add the watch until the directory
is created by the pull program.