Perform a clean shutdown in case no tasks are running
If we receive a termination signal while there are no pending tasks
we can't signal the event loop to be stopped. In this case we can
safely exit the main program only after we stop the executor in order
to make sure there aren't any pending IO operations.

This commit adds some more debug logging as it was difficult to spot
the race condition of ignoring termination signals in case of no
pending tasks in the event loop.