Shield workers from death due to errors from dispatchers
This commits implements a retry logic on Graphite dispatcher which
prevents the death of the workers when Graphite relay is unavailable.
It also includes a deque type of Queue in which we store metrics while
Graphite relay is unavailable. In order to avoid slowdowns on processing
we delay a reconnection few seconds and flush all data from the queue
when Graphite relay is available again. We also implement a simple
backoff retry logic, we should add also an exponential backoff with
jitter.

We also shield FileHandler from possible breakages but without any
queue. If we can't write to the file, the FileHandler drops the metrics.

This commits changes changes few parameters::
    retry --> retries
    + delay
    + backoff
    + queue-size
    - sent-failed