Rework the retries decorator
Used to inspect the decorated function in order to fetch parameters,
which isn't the way a decorator should work. We fix this by
introducing a wrapper function (connect) which has access to parameters
of GraphiteHandler and can passed them to the decorator at parse time.

We also add proper docstrings in few functions/methods.

There *isn't* any change in the functionality of retries decorators.