Fix a regression introduced by b61ce6bcc
In b61ce6bcc we introduced a pull-timeout which sets a timeout for
pulling data(info and stats) from all UNIX sockets.  But, we never
checked if they actual pull of data finished successfully.
The coroutines returned False as they received exception when
trying to connect to UNIX sockets, but their status wasn't checked.
As a result supervisor() was assuming everything went fine and moved the
directory with no files to final destination. The process program didn't
handle this situation very well as workers got exceptions from Pandas as
we passed empty data and crashed, actually workers became orphan
processes:-(. Process program is now more robust and avoids to pass
empty data to Pandas.

With this commit we check the result of each finished task and we claim
success only when all tasks are finished successfully.