Handle TCP socket failures in a better way
We don't handle very well TCP failures when Graphite relay is in bad
state. We also don't close previously connection in some situations.
Which causes even more troubles to Graphite relay as it reaches file
descriptor limits. 11decc9d commit made the situation worse due to the
way catching exception works. We first catch OSError and then
socket.timeout. We only close the connection at the block of
socket.timeout, but we never hit that block as OSΕrrror is parent
class of socket.timeout.

This commit fixes these problems and it adds more information when
Graphite is in bad state, which helps troubleshooting/debug.