Add connect and write timeout to defaults/README
We missed to add to the defaults the connect-timeout and write-timeout
parameters for graphite section. We also forgot to mention them in the
README.

This was *hidden* by the fact we use 'fallback' when we try to get them
from ConfigParser object, so we never got an exception from ConfigParser.