Add a missing comma
The missing comma confuses a lot the operations on Pandas frames
and as a result we send incorrect Graphite data::

...haproxy.daemon.process.1.ConnRateLimitCumConns nan 1460536614
                            ^^^^^^^^^^^^^^^^^^^^  ^^^
The issue doesn't present itself in production as carbon-c-relay
ignores those metrics and they never arrive on Graphite.

During the development of calculating percentages it was presented itself
as it caused crashed on some Pandas operations.