Calculate percentages for HAProxy workers
We calculate percentages for a selection of metrics for HAProxy daemon.
When **per-process-metrics** is set to **true** the calculation happens
also per HAProxy process. This adds the following metric names::

    ConnPercentage
    ConnRatePercentage
    SslRatePercentage

Those metrics can be used for alerting when the current usage on
connections is very close the configured limit.

This addresses PR #1 in a more generic way and it is not enabled by
default.