Do a comparison to None singleton with is not
PEP8:
Also, beware of writing if x when you really mean if x is not None --
e.g. when testing whether a variable or argument that defaults to None
was set to some other value. The other value might have a type (such as
a container) that could be false in a boolean context!

raised variable defaults to None and it may point to an Exception
when a failure occurs. So, it doesn't make sense to check if it is
True or False. Just follow PEP8 recommendation and you are safe, they
say:-)