Avoid replacing '.' with '_' in namespace setting
If namespace is set to a string which contains '.' characters
we accidentally replace them with '_' character and as a result
we send graphite metrics which aren't accepted by graphite relays.

For example with the following setting::

    [graphite]
    ...
    namespace = loadbalancers.group.internal
    ...

we send metrics like::

    loadbalancers_group_internal.poseidonas.haproxy.backend

instead of sending::

    loadbalancers.group.internal.poseidonas.haproxy.backend

We fix this by only replacing possible occurrence of '.' with  '_' in
the FQDN of the node, when prefix_hostname is set to true.