Fix performance regression introduced by 9e78b6918
Shortly after 0.3.11 version was pushed to production, we noticed a huge
increase in the 95th percentile of CPU utilization at user level. We
also noticed a huge increase ( 3 times ) of WallClockTime for processing
server stats. We rolled back to 0.3.10 and CPU utilization and WallClockTime
restored to previous levels.

After a short git bisecting session I found the bad commit,
"Sanitize values in pxname and svname columnsi".

On the development environment wasn't that easy to spot the performance
impact of that commit due to small number of frontends, backends, and servers.
After starting HAProxy with ~50 frontends and backends which had 120
servers each, I saw the spike in WallClockTime.

The root cause was the lookup for values in Pandas in a nested loop.
We fix this by storing the value in a variable before entering the inner
loop.

Here is the WallClockTime before the fix::

    wall clock time in seconds for haproxy_stats 0.041
    wall clock time in seconds for process_frontends 0.035
    wall clock time in seconds for process_backends 0.135
    wall clock time in seconds for process_servers 11.961

and here is with the fix::

    wall clock time in seconds for haproxy_stats 0.019
    wall clock time in seconds for process_frontends 0.024
    wall clock time in seconds for process_backends 0.046
    wall clock time in seconds for process_servers 2.839