Add support for pulling statistics from TCP sockets
This commit adds support for retrieving HAProxy statistics by connecting
to TCP sockets, in addition to UNIX sockets.

We introduce a new option with the name 'servers' in the 'pull' section.
The 'servers' option holds a list of servers separated by comma.
We define a server by passing a URL, here some examples::

    tcp://127.0.0.1:5555
    tcp://foo.bar.com:4444
    tcp://[fe80::3f2f:46b3:ef0c:a420]:4444
    unix:///run/haproxy.sock

We support only TCP and UNIX schemes and the port for TCP servers **must**
be set.

For UNIX scheme you can only pass a file and not a directory, but
**socket-dir** option can be set as well, so you can use a directory and
UNIX socket files at the same time.

We use comma as separator to pass multiple servers::

    servers = unix:///run/haproxy.sock,tcp://127.0.0.1:555,tcp://127.0.0.1:556

In order to allow users to set only servers and not a socket directory,
we have to drop the default value for 'socket-dir', before we used
'/run/haproxy' as default value.

We also don't have a default value for 'servers' option, so users have
the flexibility to set either only a socket dir, or only socket files,
or only TCP sockets or a combination of all the previous.

This commit introduces also a new option for 'pull' section to set a
limit for the read buffer we use when we retrieve data from UNIX/TCP
sockets. Before, we didn't have a limit.

The name of the option is 'buffer-limit' and defaults to 5MB.
If we set a very low value, then we slow down the pull of statistics.
Only values greater than or equal to 1 are accepted.