Check liveness of workers and exit if at least one is dead
We used to have a Monit check to monitor the number of alive workers
and trigger a restart if at least one was dead. This was an OKish
solution but not the best for everyone.

We now have a thread that check the liveness of the workers and
terminates ourself if at least one is dead. Systemd detects the
termination and restarts the process.