Fix runtime error when specifying a custom list of metrics
This error is triggered by specifying 'server-metrics', 'backend-metrics'
or 'frontend-metrics' in the [process] section of the config file.
The cause is an attempt to access missing variables in the global namespace.

Note this patch also fixes a secondary issue: the list of valid
metrics didn't include the AVG metrics.

Traceback (most recent call last):
  File ".../haproxystats-process", line 10, in <module>
    sys.exit(main())
  File ".../haproxystats/process.py", line 816, in main
    check_metrics(config)
  File ".../haproxystats/utils.py", line 668, in check_metrics
    if not set(valid_metrics).issuperset(metrics):
TypeError: 'NoneType' object is not iterable